README
Warning : if your emails get marked as spam by your target you might lose the email address that you used . build an email that you don't mind to lose it !
step 1) open the project in pycharm and fill the "enter your email " parts with your email . 
step 2) your excel list should be named "email.xlsx" . copy it next to the python code . they have to be in the same folder .
step 3) Edit your massage . it's an html massage . you can use some template for it. i uploaded some template next to the project as well. 
step 4) Run and enjoy. 
note : if it got crashed for network problem or anything else , there is a for loop at the program . change the value of "i" to the number of the cell list that you want to start . for example if it crashed on 247 you should start from 248 . so change the i value to 248 .
Project is done by : atefeh.khazraei@yahoo.com